(initial) Add the src code for fetching and building an up-to-date PE
This commit adds the source code for the `phusion` build tool which
fetches the latest repositories for PE and merges them in with the
current build. The user may customize the branch and github user they
want to pull the repos from so that one might be able to test another
users branch such as testing for PRs. The tool is designed to be
lightweight and fast with easy configuration. Currently there is a
dependency on Puppet (specifically the module tool), we hope to be able
to port over the builder source code to this project so that we may
eliminate this dep.